requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
python-multipart==0.0.6
openai-whisper==20231117
numpy>=1.20.0
torch>=2.0.0 